modularized bg video fetching code
